+++ RELEASENOTES.md
@@ -1,4 +1,19 @@
 ##################################################
+# order-service-3.16.0-SNAPSHOT |5-Dec-2022
+##################################################
+1. New API for request generate qrcode url for Dine-In 
+2. New API for validate qrcode token during scan the qrcode
+
+##DB Changes:
+ CREATE TABLE qrcode_session (
+ token VARCHAR(50) PRIMARY KEY,
+ tagId BIGINT,
+ created DATETIME
+ );
+ 
+ 
+ 
+##################################################
 # order-service-3.15.2-SNAPSHOT |9-Nov-2022
 ##################################################
 New API to update order shipment phone number